Modify the compiler to allow variable redeclaration with value update
This commit introduces changes to the compiler to permit the redeclaration of variables, including simple variables, structs, and arrays. Now, the compiler will accept variable redeclaration as long as they are of the same type and dimension in the case of arrays.

Additionally, functionality has been implemented to update the values of variables when assigning a value during their declaration. This ensures that when specifying a value during declaration, this value is maintained and updated in the variable definition.